// src/frontend/src/declarations/backend.did.d.ts
export type CollectionVisibility = { 'public' : null } |
  { 'private' : null };
export interface Comment {
  'id' : CommentId,
  'authorId' : UserId,
  'createdAt' : Timestamp,
  'text' : string,
  'timestampSeconds' : bigint,
  'videoId' : VideoId,
}
export type CommentId = bigint;
export interface CreateCollectionInput {
  'title' : string,
  'description' : string,
  'visibility' : CollectionVisibility,
}
export interface CreateCommentInput {
  'text' : string,
  'timestampSeconds' : bigint,
  'videoId' : VideoId,
}
export interface CreateVideoInput {
  'title' : string,
  'description' : string,
  'videoBlob' : ExternalBlob,
  'thumbnailBlob' : [] | [ExternalBlob],
  'durationSeconds' : bigint,
  'visibility' : VideoVisibility,
}
export type ExternalBlob = Uint8Array;
export type InvitationId = bigint;
export type InvitationStatus = { 'revoked' : null } |
  { 'expired' : null } |
  { 'pending' : null } |
  { 'accepted' : null } |
  { 'declined' : null };
export interface SaveProfileInput {
  'bio' : string,
  'displayName' : string,
  'profilePicture' : [] | [ExternalBlob],
}
export type Timestamp = bigint;
export interface UpdateCollectionInput {
  'title' : string,
  'description' : string,
  'visibility' : CollectionVisibility,
}
export interface UpdateVideoInput {
  'title' : string,
  'description' : string,
  'visibility' : VideoVisibility,
}
// ...
export type VideoVisibility = { 'public' : null } |
  { 'private' : null } |
  { 'collaboratorsOnly' : null };